refactor(server): fixed & removed temp items
