+++ server/.eslintignore
@@ -1,1 +1,4 @@
 /dist
+node_modules
+.node_modules
+/node_modules
