+++ server/.lintstagedrc.json
@@ -1,3 +1,3 @@
 {
-  "*.ts": ["npm run lint"]
+  "*.ts": ["yarn lint"]
 }
