--- server/docker-compose.yml
@@ -1,50 +1,0 @@
-version: '3.7'
-
-services:
-  proxy:
-    image: nginx:alpine
-    container_name: proxy
-    ports:
-      - '80:80'
-    volumes:
-      - ./nginx.conf:/etc/nginx/nginx.conf
-    restart: 'unless-stopped'
-    networks:
-      - backend
-  server:
-    build:
-      context: ./
-      target: development-build-stage
-      dockerfile: Dockerfile
-    container_name: server
-    ports:
-      - '3000:3000'
-    environment:
-      MONGO_HOST: mongo
-      MONGO_PORT: 5432
-      MONGO_DATABASE: mongoose
-    volumes:
-      - ./:/app
-      - /app/node_modules
-    restart: 'unless-stopped'
-    networks:
-      - backend
-    links:
-      - mongo
-    depends_on:
-      - mongo
-  mongo:
-    image: mvertes/alpine-mongo
-    container_name: mongo
-    ports:
-      - '5432:5432'
-    networks:
-      - backend
-
-networks:
-  backend:
-    driver: bridge
-
-volumes:
-  data:
-    driver: local
