+++ server/package.json
@@ -7,7 +7,6 @@
     "start": "npm run build && cross-env NODE_ENV=production node dist/server.js",
     "dev": "cross-env NODE_ENV=development nodemon",
     "build": "tsc && npx tsc-alias",
-    "test": "jest --forceExit --detectOpenHandles",
     "lint": "eslint --ignore-path .gitignore --ext .ts src/",
     "lint:fix": "npm run lint -- --fix",
     "deploy:prod": "npm run build && pm2 start ecosystem.config.js --only prod",
@@ -15,8 +14,6 @@
   },
   "dependencies": {
     "axios": "^0.24.0",
-    "class-transformer": "^0.4.0",
-    "class-validator": "^0.13.1",
     "compression": "^1.7.4",
     "config": "^3.3.6",
     "cookie-parser": "^1.4.5",
@@ -27,14 +24,10 @@
     "express": "^4.17.1",
     "helmet": "^4.6.0",
     "hpp": "^0.2.3",
-    "jest": "^27.0.6",
     "jsonwebtoken": "^8.5.1",
     "morgan": "^1.10.0",
     "pg": "^8.7.1",
     "pm2": "^5.1.0",
-    "swagger-jsdoc": "^6.0.0",
-    "swagger-ui-express": "^4.1.6",
-    "ts-jest": "^27.0.3",
     "ts-node": "^10.0.0",
     "typeorm": "^0.2.41",
     "typescript": "^4.3.5",
@@ -55,9 +48,6 @@
     "@types/jsonwebtoken": "^8.5.4",
     "@types/morgan": "^1.9.3",
     "@types/node": "^16.0.1",
-    "@types/supertest": "^2.0.11",
-    "@types/swagger-jsdoc": "^6.0.1",
-    "@types/swagger-ui-express": "^4.1.3",
     "@types/winston": "^2.4.4",
     "@typescript-eslint/eslint-plugin": "^4.28.2",
     "@typescript-eslint/parser": "^4.28.2",
@@ -67,7 +57,6 @@
     "node-config": "^0.0.2",
     "node-gyp": "^8.1.0",
     "nodemon": "^2.0.9",
-    "supertest": "^6.1.3",
     "tsconfig-paths": "^3.10.1"
   }
 }
