+++ server/src/app.ts
@@ -8,13 +8,11 @@ import express from 'express';
 import helmet from 'helmet';
 import hpp from 'hpp';
 import morgan from 'morgan';
-import swaggerJSDoc from 'swagger-jsdoc';
-import swaggerUi from 'swagger-ui-express';
 import { dbConnection } from '@databases';
-import { Routes } from '@interfaces/routes.interface';
 import errorMiddleware from '@/middlewares/errorMiddleware';
 import { logger, stream } from '@utils/logger';
 import { createConnection } from 'typeorm';
+import { Routes } from './interfaces';
 
 class App {
   public app: express.Application;
@@ -29,7 +27,6 @@ class App {
     this.connectToDatabase();
     this.initializeMiddlewares();
     this.initializeRoutes(routes);
-    this.initializeSwagger();
     this.initializeErrorHandling();
   }
 
@@ -67,22 +64,6 @@ class App {
     routes.forEach((route) => {
       this.app.use('/', route.router);
     });
-  }
-
-  private initializeSwagger() {
-    const options = {
-      swaggerDefinition: {
-        info: {
-          title: 'REST API',
-          version: '1.0.0',
-          description: 'Example docs',
-        },
-      },
-      apis: ['swagger.yaml'],
-    };
-
-    const specs = swaggerJSDoc(options);
-    this.app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(specs));
   }
 
   private initializeErrorHandling() {
