--- server/src/dtos/users.dto.ts
@@ -1,9 +1,0 @@
-import { IsEmail, IsString } from 'class-validator';
-
-export class CreateUserDto {
-  @IsEmail()
-  public email: string;
-
-  @IsString()
-  public password: string;
-}
