+++ server/src/interfaces.ts
@@ -1,3 +1,5 @@
+/* entities */
+
 export interface SetInterface {
   id: string;
   title: string;
@@ -20,7 +22,6 @@ export interface FolderInterface {
   updatedAt: string;
 }
 
-/* user */
 export interface UserInterface {
   id: string;
   email: string;
@@ -32,4 +33,40 @@ export interface UserInterface {
   fingerprint: string;
   createdAt: string;
   updatedAt: string;
+}
+
+/* other */
+
+export interface dbConfig {
+  host: string;
+  port: number;
+  database: string;
+  user: string;
+  password: string;
+  synchronize: boolean;
+}
+
+import { Router } from 'express';
+export interface Routes {
+  path?: string;
+  router: Router;
+}
+
+export interface GoogleUser {
+  id: string;
+  email: string;
+  verified_email: boolean;
+  name: string;
+  given_name: string;
+  family_name: string;
+  picture: string;
+  locale: string;
+}
+
+export interface GoogleTokens {
+  access_token: string;
+  expires_in: Number;
+  refresh_token: string;
+  scope: string;
+  id_token: string;
 }
