--- server/src/interfaces/db.interface.ts
@@ -1,8 +1,0 @@
-export interface dbConfig {
-  host: string;
-  port: number;
-  database: string;
-  user: string;
-  password: string;
-  synchronize: boolean;
-}
