--- server/src/interfaces/routes.interface.ts
@@ -1,6 +1,0 @@
-import { Router } from 'express';
-
-export interface Routes {
-  path?: string;
-  router: Router;
-}
