--- server/src/tests/index.test.ts
@@ -1,18 +1,0 @@
-import request from 'supertest';
-import App from '@/app';
-import IndexRoute from '@/routes/indexRoute';
-
-afterAll(async () => {
-  await new Promise<void>((resolve) => setTimeout(() => resolve(), 500));
-});
-
-describe('Testing Index', () => {
-  describe('[GET] /', () => {
-    it('response statusCode 200', () => {
-      const indexRoute = new IndexRoute();
-      const app = new App([indexRoute]);
-
-      return request(app.getServer()).get(`${indexRoute.path}`).expect(200);
-    });
-  });
-});
