+++ server/src/utils/isEmpty.ts
